+++ src/locale.js
@@ -139,11 +139,8 @@ define(function(require) {
 			};
 		},
         load: function (name, req, onLoad, config) {
-        	if(name === ".") {
-        		debugger;
-        	}
-
-        	req([(window.locale_base || "locale/") + name], function(dict) {
+        	var base = (window.locale_base || "locale/");
+        	req([base + "prototype", base + name], function(proto, dict) {
         		locale[name] = js.mixIn(js.obj2kvp(dict.proto || {}), js.obj2kvp(dict));
         		onLoad(dict);
         	});